+++ app/src/main/java/com/vp/budgetmanager/database/TransactionRepository.kt
@@ -7,11 +7,11 @@ import kotlinx.coroutines.flow.Flow
 class TransactionRepository(private val transactionDAO: TransactionDAO) {
 
     val allTransactions: Flow<List<Transaction>> = transactionDAO.getAllTransactions()
-    val totalSpent: Flow<Int> = transactionDAO.getTotalSpent()
+    val totalSpent: Flow<Float> = transactionDAO.getTotalSpent()
 
     @Suppress("RedundantSuspendModifier")
     @WorkerThread
-    fun getTodaySpent(time: Long): Flow<Int>{
+    fun getTodaySpent(time: Long): Flow<Float>{
         return transactionDAO.getTodaySpent(time)
     }
